# bot_dialog_generator/contracts/openapi/studio-api.yaml
openapi: 3.1.0
info:
  title: Bot_Dialog_Generator Studio API
  description: |
    Authoring, draft versioning and publishing API for Bot_Dialog_Generator bots.

    Every bot belongs to a workspace (one per area of the organization). All workspace
    routes require a bearer token. Callers who are not members of a workspace receive
    404 for anything inside it, so its existence is not revealed (ADR 0003). Platform
    administrators can access every workspace; each such access is written to the audit log.

    Roles inside a workspace: `analyst` (read), `editor` (read, create bots, edit drafts),
    `owner` (everything, including publish, rollback and managing members).

    Access comes from direct memberships and from identity-provider groups. With
    Microsoft Entra ID, a group is identified by its object ID, which is what the token's
    `groups` claim contains. A workspace always keeps at least one owner. Archived
    workspaces are hidden from their members; platform admins can restore them.
  version: 0.3.0
security:
  - bearerAuth: []
paths:
...
  /dev/token:
    post:
      summary: Mint a development token
      description: |
        Available only when the service runs with AUTH_MODE=dev, which is refused outside
        local and dev environments. Mints a signed token for any subject and groups.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevTokenRequest'
      responses:
        '200':
          description: Signed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /me:
    get:
      summary: Current caller
      description: The authenticated subject, whether it is a platform admin, and the workspaces it belongs to.
      responses:
        '200':
          description: Caller identity and memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /workspaces:
    get:
      summary: List my workspaces
      description: Only workspaces the caller is a member of, directly or through an identity-provider group.
      responses:
        '200':
          description: Workspaces with the caller's role in each
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceAccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /workspaces/{workspaceId}/bots:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      summary: List bots in a workspace
      description: Requires the analyst role.
      responses:
        '200':
          description: Bots in the workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create bot
      description: Requires the editor role. Creates the bot with a starter draft `v1` containing one Trigger node.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotRequest'
      responses:
        '201':
          description: Bot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/bots/{botId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BotId'
    get:
      summary: Get bot
      description: Requires the analyst role.
      responses:
        '200':
          description: Bot including its active and draft versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /workspaces/{workspaceId}/bots/{botId}/versions:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BotId'
    get:
      summary: List bot versions
      description: Requires the analyst role. Ordered by creation time.
      responses:
        '200':
          description: Versions of the bot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Version'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create draft version
      description: Requires the editor role. Copies nodes and edges from `base_version` when given.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionRequest'
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/bots/{botId}/versions/{versionId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BotId'
      - $ref: '#/components/parameters/VersionId'
    get:
      summary: Get version
      description: Requires the analyst role.
      responses:
        '200':
          description: Full version with nodes and edges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Save draft
      description: Requires the editor role. Replaces the draft's entry node, nodes and edges. Published versions are immutable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveDraftRequest'
      responses:
        '200':
          description: Saved draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '400':
          description: Invalid request, or the version is already published
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/bots/{botId}/versions/{versionId}/publish:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BotId'
      - $ref: '#/components/parameters/VersionId'
    post:
      summary: Publish draft
      description: |
        Requires the owner role. Validates the draft, compiles the canonical protobuf
        definition, stores it under the workspace's prefix, moves the active version
        pointer and emits a `bot.published` outbox event. Audited.
      responses:
        '200':
          description: Published version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResult'
        '400':
          description: Draft validation failed, or the version is already published
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/members:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      summary: List who has access to the workspace
      description: Requires the analyst role.
      responses:
        '200':
          description: People and identity-provider groups with a role in this workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Access'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/members/{subject}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/Subject'
    put:
      summary: Add a person or change their role
      description: Requires the owner role. Fails with 400 if it would leave the workspace without an owner. Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: The membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Remove a person
      description: Requires the owner role. Removing someone who is not a member succeeds. Fails with 400 if it would leave the workspace without an owner. Audited.
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/GroupId'
    put:
      summary: Give an identity-provider group a role
      description: Requires the owner role. Every member of the group gets the role. Fails with 400 if it would leave the workspace without an owner. Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                display_name:
                  type: string
                  maxLength: 100
                  description: Name shown in the studio; defaults to the group ID.
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: The group grant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Remove a group's access
      description: Requires the owner role. Removing a group without access succeeds. Fails with 400 if it would leave the workspace without an owner. Audited.
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/workspaces:
    get:
      summary: List all workspaces
      description: Platform admins only, including archived workspaces. Audited.
      responses:
        '200':
          description: Every workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create a workspace (an area)
      description: Platform admins only. The workspace needs at least one owner, a person or a group. Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: The new workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    patch:
      summary: Rename, archive or restore a workspace
      description: Platform admins only. Omitted fields are unchanged. Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: The updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/bots:
    get:
      summary: List all bots in every workspace
      description: Platform admins only. Audited.
      responses:
        '200':
          description: Every bot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/audit:
    get:
      summary: Read the audit log
      description: Platform admins only. Audited.
      responses:
        '200':
          description: Audit entries, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC token from the organization's identity provider (or from /dev/token in local development).
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Missing, invalid or expired bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    Me:
      type: object
      required: [subject, platform_admin, workspaces]
      properties:
        subject: { type: string }
        email: { type: string }
        platform_admin: { type: boolean }
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceAccess'
    DevTokenRequest:
      type: object
      required: [subject]
      properties:
        subject: { type: string, minLength: 1 }
        groups:
          type: array
          items: { type: string }
    DevTokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
...